backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]
sqlalchemy
pymysql
python-multipart
argon2-cffi
passlib[argon2]
python-jose[cryptography]
alembic
pydantic==2.7.4
pydantic-settings==2.2.1
PyYAML==6.0.1
httpx==0.27.0
